Move header into separate folder
